Mocked object can be created using parameters.
Since TestRule is supposed to modify a single test but not the testing
mechnisim, adding parameters for tests is not supported by TestRule.
Since the purpose of TestRule is providing isolated additional function to
tests, the HiMockTestRule will not be updated from now. But if there will be
features that can be isolated from the whole system, more rules will be
added.